requirements.txt
PyYAML
rich
requests
